Move org vault into organization-vault module
