+++ src/app/modules/vault/modules/organization-vault/collections.component.ts
@@ -11,11 +11,11 @@ import { Cipher } from "jslib-common/models/domain/cipher";
 import { Organization } from "jslib-common/models/domain/organization";
 import { CipherCollectionsRequest } from "jslib-common/models/request/cipherCollectionsRequest";
 
-import { CollectionsComponent as BaseCollectionsComponent } from "../../vault/collections.component";
+import { CollectionsComponent as BaseCollectionsComponent } from "../../../../vault/collections.component";
 
 @Component({
   selector: "app-org-vault-collections",
-  templateUrl: "../../vault/collections.component.html",
+  templateUrl: "../../../../vault/collections.component.html",
 })
 export class CollectionsComponent extends BaseCollectionsComponent {
   organization: Organization;
